+++ Source/Blazorise/Base/BaseBarDropdownToggle.cs
@@ -9,15 +9,11 @@ using Microsoft.AspNetCore.Blazor.Components;
 
 namespace Blazorise.Base
 {
-    public abstract class BaseDropdownToggle : BaseComponent, ICloseActivator
+    public abstract class BaseBarDropdownToggle : BaseComponent, ICloseActivator
     {
         #region Members
 
         private bool isOpen;
-
-        private bool actAsButton;
-
-        private bool isSplit;
 
         private bool isRegistered;
 
@@ -44,8 +40,7 @@ namespace Blazorise.Base
         protected override void RegisterClasses()
         {
             ClassMapper
-                .If( () => ClassProvider.DropdownToggle(), () => !ActAsButton )
-                .If( () => ClassProvider.DropdownToggleSplit(), () => !ActAsButton && IsSplit );
+                .Add( () => ClassProvider.BarDropdownToggle() );
 
             base.RegisterClasses();
         }
@@ -53,15 +48,14 @@ namespace Blazorise.Base
         protected override void OnInit()
         {
             // link to the parent component
-            Dropdown?.Hook( this );
+            BarDropdown?.Hook( this );
 
             base.OnInit();
         }
 
-
         protected void ClickHandler()
         {
-            Dropdown?.Toggle();
+            BarDropdown?.Toggle();
         }
 
         public bool SafeToClose( string elementId )
@@ -71,27 +65,12 @@ namespace Blazorise.Base
 
         public void Close()
         {
-            Dropdown?.Close();
+            BarDropdown?.Close();
         }
 
         #endregion
 
         #region Properties
-
-        /// <summary>
-        /// Occurs when the toggle button is clicked.
-        /// </summary>
-        [Parameter] protected Action<bool> Toggle { get; set; }
-
-        /// <summary>
-        /// Gets or sets the dropdown color.
-        /// </summary>
-        [Parameter] protected Color Color { get; set; } = Color.None;
-
-        /// <summary>
-        /// Gets or sets the dropdown size.
-        /// </summary>
-        [Parameter] protected Size Size { get; set; } = Size.None;
 
         /// <summary>
         /// Handles the visibility of dropdown toggle.
@@ -121,42 +100,7 @@ namespace Blazorise.Base
             }
         }
 
-        /// <summary>
-        /// Button outline.
-        /// </summary>
-        [Parameter] protected bool IsOutline { get; set; }
-
-        /// <summary>
-        /// Act as ordinary button instead of dropdown toggle.
-        /// </summary>
-        [Parameter]
-        protected bool ActAsButton
-        {
-            get => actAsButton;
-            set
-            {
-                actAsButton = value;
-
-                ClassMapper.Dirty();
-            }
-        }
-
-        /// <summary>
-        /// Handles the visibility of split button.
-        /// </summary>
-        [Parameter]
-        protected bool IsSplit
-        {
-            get => isSplit;
-            set
-            {
-                isSplit = value;
-
-                ClassMapper.Dirty();
-            }
-        }
-
-        [CascadingParameter] protected Dropdown Dropdown { get; set; }
+        [CascadingParameter] protected BaseBarDropdown BarDropdown { get; set; }
 
         [Parameter] protected RenderFragment ChildContent { get; set; }